Optimize subscribe performance
IO#gets(delimiter : String) is slow, but IO#gets(delimiter : Char) is
faster. I was able to achieve a ~43% improvement in subscriber
throughput (1.86M vs 1.27M) by using a Char instead of a String
delimeter.

This appears to be safe since there isn't anywhere that a \n could
appear in either of these cases where a \r wouldn't precede it. So as
long as the server is doing the right thing, we should be good.